requirements.txt
fastapi
uvicorn[standard]
duckdb
mangum
python-dotenv
supabase
google-generativeai
upstash-redis
pydantic
